requirements.txt
pycryptodome
PyPDF2
reportlab
